Basic error checking on attempted test.
- Prints attempted command to log.
- Also prints if an exception occured.
- Do not clean up temp files from a failed test.

+++ test_integration/geopm_test_integration.py
@@ -259,7 +259,10 @@ class Launcher(object):
         env.update(self._environ())
         self._app_conf.write()
         self._ctl_conf.write()
+        # The following open() is required to preserve the proper output ordering when stdout and stderr are combined.
         with open(test_name + '.log', 'w') as outfile:
+            outfile.write(str(self) + '\n\n')
+            outfile.flush()
             subprocess.check_call(self._exec_str(), shell=True, env=env, stdout=outfile, stderr=outfile)
 
     def get_report(self):
@@ -348,13 +351,15 @@ class TestReport(unittest.TestCase):
                          'power_budget' : 150}
         self._epsilon = 0.05
         self._tmp_files = []
+        self._do_cleanup = False
 
     def tearDown(self):
-        for ff in self._tmp_files:
-            try:
-                os.remove(ff)
-            except OSError:
-                pass
+        if self._do_cleanup:
+            for ff in self._tmp_files:
+                try:
+                    os.remove(ff)
+                except OSError:
+                    pass
 
     def assertNear(self, a, b):
         if abs(a - b) / a >= self._epsilon:
@@ -380,6 +385,7 @@ class TestReport(unittest.TestCase):
         for ff in reports:
             self.assertTrue(os.path.isfile(ff))
             self.assertTrue(os.stat(ff).st_size != 0)
+        self._do_cleanup = True
 
     def test_runtime(self):
         name = 'test_runtime'
@@ -403,6 +409,7 @@ class TestReport(unittest.TestCase):
         for rr in reports:
             self.assertNear(delay, rr['sleep'].get_runtime())
             self.assertGreater(rr.get_runtime(), rr['sleep'].get_runtime())
+        self._do_cleanup = True
 
     def test_progress(self):
         name = 'test_progress'
@@ -426,6 +433,7 @@ class TestReport(unittest.TestCase):
         for rr in reports:
             self.assertNear(delay, rr['sleep'].get_runtime())
             self.assertGreater(rr.get_runtime(), rr['sleep'].get_runtime())
+        self._do_cleanup = True
 
 if __name__ == '__main__':
     unittest.main()
